Refactor to use ES6 notation
