+++ index.html
@@ -160,7 +160,7 @@
           'We use cookies to improve your experience. By your continued use of this site you accept such use.<br /> This notice will disappear by itself.',
         duration: 20000,
       };
-      cpNs.cookiePolicy.setup(options);
+      cpNs.setup(options);
     </script>
   </body>
 </html>
